pyproject.toml
[project]
name = "gfog"
version = "0.1.0"
description = "Gfog: Gradient free optimization via gradients"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "attrs>=25.3.0",
  "debugpy>=1.8.14",
  "imageio>=2.37.0",
  "loguru>=0.7.3",
  "matplotlib>=3.10.3",
  "numpy>=2.3.1",
  "rich>=14.0.0",
  "seaborn>=0.13.2",
  "sortedcontainers>=2.4.0",
  "torch>=2.7.1",
  "torchvision>=0.22.1",
  "tqdm>=4.67.1",
]
